+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# [4.1.0-alpha.6](https://github.com/ar-io/ar-io-sdk/compare/v4.1.0-alpha.5...v4.1.0-alpha.6) (2026-07-18)
+
+
+### Bug Fixes
+
+* **solana:** don't let sync_attributes revert the spawned-ANT ACL records ([ba3cd1e](https://github.com/ar-io/ar-io-sdk/commit/ba3cd1e415d0ff712a569e52a9a36765349d21cf))
+* **solana:** route oversized atomic spawn-and-buy through a lookup table ([a929556](https://github.com/ar-io/ar-io-sdk/commit/a92955603d9b34951b5073f25b35aaa3aaf47040))
+
+# [4.1.0-alpha.5](https://github.com/ar-io/ar-io-sdk/compare/v4.1.0-alpha.4...v4.1.0-alpha.5) (2026-07-17)
+
+
+### Features
+
+* **solana:** atomically spawn an ANT during buyRecord when no processId ([35b3dea](https://github.com/ar-io/ar-io-sdk/commit/35b3dea306b6bc3461a25122581f4c13eca07ff0))
+
 # [4.1.0-alpha.4](https://github.com/ar-io/ar-io-sdk/compare/v4.1.0-alpha.3...v4.1.0-alpha.4) (2026-07-02)